+++ test/gencode/lib/testdb.c
@@ -9,7 +9,7 @@
 #ifdef TESTDB_USE_DIAG_MONITORS
 // Function prototypes to be called each time CAN frame is unpacked
 // FMon function may detect RC, CRC or DLC violation
-#include <testdb-fmon.h>
+#include "testdb-fmon.h"
 
 #endif // TESTDB_USE_DIAG_MONITORS
 
@@ -47,12 +47,12 @@ static bitext_t __ext_sig__(ubitext_t val, uint8_t bits)
 uint32_t Unpack_UTEST_2_testdb(UTEST_2_t* _m, const uint8_t* _d, uint8_t dlc_)
 {
   (void)dlc_;
-  _m->U28_TEST_1 = ((_d[3] & (0x0FU)) << 24) | ((_d[2] & (0xFFU)) << 16) | ((_d[1] & (0xFFU)) << 8) | (_d[0] & (0xFFU));
-  _m->ValTest = ((_d[3] >> 5) & (0x03U));
-  _m->U8_TEST_1 = (_d[4] & (0xFFU));
-  _m->U7_TEST_1_ro = ((_d[5] >> 1) & (0x7FU));
+  _m->U28_TEST_1 = (uint32_t) ( ((_d[3] & (0x0FU)) << 24U) | ((_d[2] & (0xFFU)) << 16U) | ((_d[1] & (0xFFU)) << 8U) | (_d[0] & (0xFFU)) );
+  _m->ValTest = (uint8_t) ( ((_d[3] >> 5U) & (0x03U)) );
+  _m->U8_TEST_1 = (uint8_t) ( (_d[4] & (0xFFU)) );
+  _m->U7_TEST_1_ro = (uint8_t) ( ((_d[5] >> 1U) & (0x7FU)) );
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->U7_TEST_1_phys = TESTDB_U7_TEST_1_ro_fromS(_m->U7_TEST_1_ro);
+  _m->U7_TEST_1_phys = (int16_t) TESTDB_U7_TEST_1_ro_fromS(_m->U7_TEST_1_ro);
 #endif // TESTDB_USE_SIGFLOAT
 
 #ifdef TESTDB_USE_DIAG_MONITORS
@@ -70,22 +70,22 @@ uint32_t Unpack_UTEST_2_testdb(UTEST_2_t* _m, const uint8_t* _d, uint8_t dlc_)
 
 uint32_t Pack_UTEST_2_testdb(UTEST_2_t* _m, __CoderDbcCanFrame_t__* cframe)
 {
-  uint8_t i; for (i = 0; (i < UTEST_2_DLC) && (i < 8); cframe->Data[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(UTEST_2_DLC); cframe->Data[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->U7_TEST_1_ro = TESTDB_U7_TEST_1_ro_toS(_m->U7_TEST_1_phys);
+  _m->U7_TEST_1_ro = (uint8_t) TESTDB_U7_TEST_1_ro_toS(_m->U7_TEST_1_phys);
 #endif // TESTDB_USE_SIGFLOAT
 
-  cframe->Data[0] |= (_m->U28_TEST_1 & (0xFFU));
-  cframe->Data[1] |= ((_m->U28_TEST_1 >> 8) & (0xFFU));
-  cframe->Data[2] |= ((_m->U28_TEST_1 >> 16) & (0xFFU));
-  cframe->Data[3] |= ((_m->U28_TEST_1 >> 24) & (0x0FU)) | ((_m->ValTest & (0x03U)) << 5);
-  cframe->Data[4] |= (_m->U8_TEST_1 & (0xFFU));
-  cframe->Data[5] |= ((_m->U7_TEST_1_ro & (0x7FU)) << 1);
+  cframe->Data[0] |= (uint8_t) ( (_m->U28_TEST_1 & (0xFFU)) );
+  cframe->Data[1] |= (uint8_t) ( ((_m->U28_TEST_1 >> 8U) & (0xFFU)) );
+  cframe->Data[2] |= (uint8_t) ( ((_m->U28_TEST_1 >> 16U) & (0xFFU)) );
+  cframe->Data[3] |= (uint8_t) ( ((_m->U28_TEST_1 >> 24U) & (0x0FU)) | ((_m->ValTest & (0x03U)) << 5U) );
+  cframe->Data[4] |= (uint8_t) ( (_m->U8_TEST_1 & (0xFFU)) );
+  cframe->Data[5] |= (uint8_t) ( ((_m->U7_TEST_1_ro & (0x7FU)) << 1U) );
 
-  cframe->MsgId = UTEST_2_CANID;
-  cframe->DLC = UTEST_2_DLC;
-  cframe->IDE = UTEST_2_IDE;
+  cframe->MsgId = (uint32_t) UTEST_2_CANID;
+  cframe->DLC = (uint8_t) UTEST_2_DLC;
+  cframe->IDE = (uint8_t) UTEST_2_IDE;
   return UTEST_2_CANID;
 }
 
@@ -93,21 +93,21 @@ uint32_t Pack_UTEST_2_testdb(UTEST_2_t* _m, __CoderDbcCanFrame_t__* cframe)
 
 uint32_t Pack_UTEST_2_testdb(UTEST_2_t* _m, uint8_t* _d, uint8_t* _len, uint8_t* _ide)
 {
-  uint8_t i; for (i = 0; (i < UTEST_2_DLC) && (i < 8); _d[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(UTEST_2_DLC); _d[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->U7_TEST_1_ro = TESTDB_U7_TEST_1_ro_toS(_m->U7_TEST_1_phys);
+  _m->U7_TEST_1_ro = (uint8_t) TESTDB_U7_TEST_1_ro_toS(_m->U7_TEST_1_phys);
 #endif // TESTDB_USE_SIGFLOAT
 
-  _d[0] |= (_m->U28_TEST_1 & (0xFFU));
-  _d[1] |= ((_m->U28_TEST_1 >> 8) & (0xFFU));
-  _d[2] |= ((_m->U28_TEST_1 >> 16) & (0xFFU));
-  _d[3] |= ((_m->U28_TEST_1 >> 24) & (0x0FU)) | ((_m->ValTest & (0x03U)) << 5);
-  _d[4] |= (_m->U8_TEST_1 & (0xFFU));
-  _d[5] |= ((_m->U7_TEST_1_ro & (0x7FU)) << 1);
+  _d[0] |= (uint8_t) ( (_m->U28_TEST_1 & (0xFFU)) );
+  _d[1] |= (uint8_t) ( ((_m->U28_TEST_1 >> 8U) & (0xFFU)) );
+  _d[2] |= (uint8_t) ( ((_m->U28_TEST_1 >> 16U) & (0xFFU)) );
+  _d[3] |= (uint8_t) ( ((_m->U28_TEST_1 >> 24U) & (0x0FU)) | ((_m->ValTest & (0x03U)) << 5U) );
+  _d[4] |= (uint8_t) ( (_m->U8_TEST_1 & (0xFFU)) );
+  _d[5] |= (uint8_t) ( ((_m->U7_TEST_1_ro & (0x7FU)) << 1U) );
 
-  *_len = UTEST_2_DLC;
-  *_ide = UTEST_2_IDE;
+  *_len = (uint8_t) UTEST_2_DLC;
+  *_ide = (uint8_t) UTEST_2_IDE;
   return UTEST_2_CANID;
 }
 
@@ -116,8 +116,8 @@ uint32_t Pack_UTEST_2_testdb(UTEST_2_t* _m, uint8_t* _d, uint8_t* _len, uint8_t*
 uint32_t Unpack_EMPTY_0_testdb(EMPTY_0_t* _m, const uint8_t* _d, uint8_t dlc_)
 {
   (void)dlc_;
-  _m->CS = ((_d[4] >> 1) & (0x3FU));
-  _m->RC = ((_d[6] & (0x07U)) << 1) | ((_d[5] >> 7) & (0x01U));
+  _m->CS = (uint8_t) ( ((_d[4] >> 1U) & (0x3FU)) );
+  _m->RC = (uint8_t) ( ((_d[6] & (0x07U)) << 1U) | ((_d[5] >> 7U) & (0x01U)) );
 
 #ifdef TESTDB_USE_DIAG_MONITORS
   _m->mon1.dlc_error = (dlc_ < EMPTY_0_DLC);
@@ -143,28 +143,28 @@ uint32_t Unpack_EMPTY_0_testdb(EMPTY_0_t* _m, const uint8_t* _d, uint8_t dlc_)
 
 uint32_t Pack_EMPTY_0_testdb(EMPTY_0_t* _m, __CoderDbcCanFrame_t__* cframe)
 {
-  uint8_t i; for (i = 0; (i < EMPTY_0_DLC) && (i < 8); cframe->Data[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(EMPTY_0_DLC); cframe->Data[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_AUTO_ROLL
   _m->RC = (_m->RC + 1) & (0x0FU);
 #endif // TESTDB_AUTO_ROLL
 
 #ifdef TESTDB_AUTO_CSM
-  _m->CS = 0U;
+  _m->CS = (uint8_t) 0;
 #endif // TESTDB_AUTO_CSM
 
-  cframe->Data[4] |= ((_m->CS & (0x3FU)) << 1);
-  cframe->Data[5] |= ((_m->RC & (0x01U)) << 7);
-  cframe->Data[6] |= ((_m->RC >> 1) & (0x07U));
+  cframe->Data[4] |= (uint8_t) ( ((_m->CS & (0x3FU)) << 1U) );
+  cframe->Data[5] |= (uint8_t) ( ((_m->RC & (0x01U)) << 7U) );
+  cframe->Data[6] |= (uint8_t) ( ((_m->RC >> 1U) & (0x07U)) );
 
 #ifdef TESTDB_AUTO_CSM
   _m->CS = ((uint8_t)GetFrameHash(cframe->Data, EMPTY_0_DLC, EMPTY_0_CANID, kXor8, 1));
-  cframe->Data[4] |= ((_m->CS & (0x3FU)) << 1);
+  cframe->Data[4] |= (uint8_t) ( ((_m->CS & (0x3FU)) << 1U) );
 #endif // TESTDB_AUTO_CSM
 
-  cframe->MsgId = EMPTY_0_CANID;
-  cframe->DLC = EMPTY_0_DLC;
-  cframe->IDE = EMPTY_0_IDE;
+  cframe->MsgId = (uint32_t) EMPTY_0_CANID;
+  cframe->DLC = (uint8_t) EMPTY_0_DLC;
+  cframe->IDE = (uint8_t) EMPTY_0_IDE;
   return EMPTY_0_CANID;
 }
 
@@ -172,27 +172,27 @@ uint32_t Pack_EMPTY_0_testdb(EMPTY_0_t* _m, __CoderDbcCanFrame_t__* cframe)
 
 uint32_t Pack_EMPTY_0_testdb(EMPTY_0_t* _m, uint8_t* _d, uint8_t* _len, uint8_t* _ide)
 {
-  uint8_t i; for (i = 0; (i < EMPTY_0_DLC) && (i < 8); _d[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(EMPTY_0_DLC); _d[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_AUTO_ROLL
   _m->RC = (_m->RC + 1) & (0x0FU);
 #endif // TESTDB_AUTO_ROLL
 
 #ifdef TESTDB_AUTO_CSM
-  _m->CS = 0U;
+  _m->CS = (uint8_t) 0;
 #endif // TESTDB_AUTO_CSM
 
-  _d[4] |= ((_m->CS & (0x3FU)) << 1);
-  _d[5] |= ((_m->RC & (0x01U)) << 7);
-  _d[6] |= ((_m->RC >> 1) & (0x07U));
+  _d[4] |= (uint8_t) ( ((_m->CS & (0x3FU)) << 1U) );
+  _d[5] |= (uint8_t) ( ((_m->RC & (0x01U)) << 7U) );
+  _d[6] |= (uint8_t) ( ((_m->RC >> 1U) & (0x07U)) );
 
 #ifdef TESTDB_AUTO_CSM
   _m->CS = ((uint8_t)GetFrameHash(_d, EMPTY_0_DLC, EMPTY_0_CANID, kXor8, 1));
-  _d[4] |= ((_m->CS & (0x3FU)) << 1);
+  _d[4] |= (uint8_t) ( ((_m->CS & (0x3FU)) << 1U) );
 #endif // TESTDB_AUTO_CSM
 
-  *_len = EMPTY_0_DLC;
-  *_ide = EMPTY_0_IDE;
+  *_len = (uint8_t) EMPTY_0_DLC;
+  *_ide = (uint8_t) EMPTY_0_IDE;
   return EMPTY_0_CANID;
 }
 
@@ -201,8 +201,8 @@ uint32_t Pack_EMPTY_0_testdb(EMPTY_0_t* _m, uint8_t* _d, uint8_t* _len, uint8_t*
 uint32_t Unpack_UTEST_3_testdb(UTEST_3_t* _m, const uint8_t* _d, uint8_t dlc_)
 {
   (void)dlc_;
-  _m->U32_TEST_1 = ((_d[3] & (0xFFU)) << 24) | ((_d[2] & (0xFFU)) << 16) | ((_d[1] & (0xFFU)) << 8) | (_d[0] & (0xFFU));
-  _m->TestValTableID = (_d[4] & (0x07U));
+  _m->U32_TEST_1 = (uint32_t) ( ((_d[3] & (0xFFU)) << 24U) | ((_d[2] & (0xFFU)) << 16U) | ((_d[1] & (0xFFU)) << 8U) | (_d[0] & (0xFFU)) );
+  _m->TestValTableID = (uint8_t) ( (_d[4] & (0x07U)) );
 
 #ifdef TESTDB_USE_DIAG_MONITORS
   _m->mon1.dlc_error = (dlc_ < UTEST_3_DLC);
@@ -219,17 +219,17 @@ uint32_t Unpack_UTEST_3_testdb(UTEST_3_t* _m, const uint8_t* _d, uint8_t dlc_)
 
 uint32_t Pack_UTEST_3_testdb(UTEST_3_t* _m, __CoderDbcCanFrame_t__* cframe)
 {
-  uint8_t i; for (i = 0; (i < UTEST_3_DLC) && (i < 8); cframe->Data[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(UTEST_3_DLC); cframe->Data[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
-  cframe->Data[0] |= (_m->U32_TEST_1 & (0xFFU));
-  cframe->Data[1] |= ((_m->U32_TEST_1 >> 8) & (0xFFU));
-  cframe->Data[2] |= ((_m->U32_TEST_1 >> 16) & (0xFFU));
-  cframe->Data[3] |= ((_m->U32_TEST_1 >> 24) & (0xFFU));
-  cframe->Data[4] |= (_m->TestValTableID & (0x07U));
+  cframe->Data[0] |= (uint8_t) ( (_m->U32_TEST_1 & (0xFFU)) );
+  cframe->Data[1] |= (uint8_t) ( ((_m->U32_TEST_1 >> 8U) & (0xFFU)) );
+  cframe->Data[2] |= (uint8_t) ( ((_m->U32_TEST_1 >> 16U) & (0xFFU)) );
+  cframe->Data[3] |= (uint8_t) ( ((_m->U32_TEST_1 >> 24U) & (0xFFU)) );
+  cframe->Data[4] |= (uint8_t) ( (_m->TestValTableID & (0x07U)) );
 
-  cframe->MsgId = UTEST_3_CANID;
-  cframe->DLC = UTEST_3_DLC;
-  cframe->IDE = UTEST_3_IDE;
+  cframe->MsgId = (uint32_t) UTEST_3_CANID;
+  cframe->DLC = (uint8_t) UTEST_3_DLC;
+  cframe->IDE = (uint8_t) UTEST_3_IDE;
   return UTEST_3_CANID;
 }
 
@@ -237,16 +237,16 @@ uint32_t Pack_UTEST_3_testdb(UTEST_3_t* _m, __CoderDbcCanFrame_t__* cframe)
 
 uint32_t Pack_UTEST_3_testdb(UTEST_3_t* _m, uint8_t* _d, uint8_t* _len, uint8_t* _ide)
 {
-  uint8_t i; for (i = 0; (i < UTEST_3_DLC) && (i < 8); _d[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(UTEST_3_DLC); _d[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
-  _d[0] |= (_m->U32_TEST_1 & (0xFFU));
-  _d[1] |= ((_m->U32_TEST_1 >> 8) & (0xFFU));
-  _d[2] |= ((_m->U32_TEST_1 >> 16) & (0xFFU));
-  _d[3] |= ((_m->U32_TEST_1 >> 24) & (0xFFU));
-  _d[4] |= (_m->TestValTableID & (0x07U));
+  _d[0] |= (uint8_t) ( (_m->U32_TEST_1 & (0xFFU)) );
+  _d[1] |= (uint8_t) ( ((_m->U32_TEST_1 >> 8U) & (0xFFU)) );
+  _d[2] |= (uint8_t) ( ((_m->U32_TEST_1 >> 16U) & (0xFFU)) );
+  _d[3] |= (uint8_t) ( ((_m->U32_TEST_1 >> 24U) & (0xFFU)) );
+  _d[4] |= (uint8_t) ( (_m->TestValTableID & (0x07U)) );
 
-  *_len = UTEST_3_DLC;
-  *_ide = UTEST_3_IDE;
+  *_len = (uint8_t) UTEST_3_DLC;
+  *_ide = (uint8_t) UTEST_3_IDE;
   return UTEST_3_CANID;
 }
 
@@ -255,38 +255,38 @@ uint32_t Pack_UTEST_3_testdb(UTEST_3_t* _m, uint8_t* _d, uint8_t* _len, uint8_t*
 uint32_t Unpack_FLT_TEST_1_testdb(FLT_TEST_1_t* _m, const uint8_t* _d, uint8_t dlc_)
 {
   (void)dlc_;
-  _m->ValTable = (_d[0] & (0x03U));
-  _m->Position = ((_d[0] >> 4) & (0x0FU));
-  _m->INT_TEST_2_ro = __ext_sig__(( ((_d[1] >> 1) & (0x7FU)) ), 7);
+  _m->ValTable = (uint8_t) ( (_d[0] & (0x03U)) );
+  _m->Position = (uint8_t) ( ((_d[0] >> 4U) & (0x0FU)) );
+  _m->INT_TEST_2_ro = (int8_t) __ext_sig__(( ((_d[1] >> 1U) & (0x7FU)) ), 7);
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->INT_TEST_2_phys = TESTDB_INT_TEST_2_ro_fromS(_m->INT_TEST_2_ro);
+  _m->INT_TEST_2_phys = (int16_t) TESTDB_INT_TEST_2_ro_fromS(_m->INT_TEST_2_ro);
 #endif // TESTDB_USE_SIGFLOAT
 
-  _m->RC = (_d[2] & (0x0FU));
-  _m->CS = ((_d[2] >> 4) & (0x0FU));
-  _m->Accel_ro = ((_d[4] & (0x0FU)) << 8) | (_d[3] & (0xFFU));
+  _m->RC = (uint8_t) ( (_d[2] & (0x0FU)) );
+  _m->CS = (uint8_t) ( ((_d[2] >> 4U) & (0x0FU)) );
+  _m->Accel_ro = (uint16_t) ( ((_d[4] & (0x0FU)) << 8U) | (_d[3] & (0xFFU)) );
 #ifdef TESTDB_USE_SIGFLOAT
   _m->Accel_phys = (sigfloat_t)(TESTDB_Accel_ro_fromS(_m->Accel_ro));
 #endif // TESTDB_USE_SIGFLOAT
 
-  _m->FLT4_TEST_1_ro = ((_d[4] >> 4) & (0x0FU));
+  _m->FLT4_TEST_1_ro = (uint8_t) ( ((_d[4] >> 4U) & (0x0FU)) );
 #ifdef TESTDB_USE_SIGFLOAT
   _m->FLT4_TEST_1_phys = (sigfloat_t)(TESTDB_FLT4_TEST_1_ro_fromS(_m->FLT4_TEST_1_ro));
 #endif // TESTDB_USE_SIGFLOAT
 
-  _m->FLT4_TEST_2_ro = (_d[5] & (0x0FU));
+  _m->FLT4_TEST_2_ro = (uint8_t) ( (_d[5] & (0x0FU)) );
 #ifdef TESTDB_USE_SIGFLOAT
   _m->FLT4_TEST_2_phys = (sigfloat_t)(TESTDB_FLT4_TEST_2_ro_fromS(_m->FLT4_TEST_2_ro));
 #endif // TESTDB_USE_SIGFLOAT
 
-  _m->FLT4_TEST_3_ro = ((_d[5] >> 4) & (0x0FU));
+  _m->FLT4_TEST_3_ro = (uint8_t) ( ((_d[5] >> 4U) & (0x0FU)) );
 #ifdef TESTDB_USE_SIGFLOAT
   _m->FLT4_TEST_3_phys = (sigfloat_t)(TESTDB_FLT4_TEST_3_ro_fromS(_m->FLT4_TEST_3_ro));
 #endif // TESTDB_USE_SIGFLOAT
 
-  _m->INT_TEST_1_ro = ((_d[6] >> 2) & (0x0FU));
+  _m->INT_TEST_1_ro = (uint8_t) ( ((_d[6] >> 2U) & (0x0FU)) );
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->INT_TEST_1_phys = TESTDB_INT_TEST_1_ro_fromS(_m->INT_TEST_1_ro);
+  _m->INT_TEST_1_phys = (int8_t) TESTDB_INT_TEST_1_ro_fromS(_m->INT_TEST_1_ro);
 #endif // TESTDB_USE_SIGFLOAT
 
 #ifdef TESTDB_USE_DIAG_MONITORS
@@ -313,41 +313,41 @@ uint32_t Unpack_FLT_TEST_1_testdb(FLT_TEST_1_t* _m, const uint8_t* _d, uint8_t d
 
 uint32_t Pack_FLT_TEST_1_testdb(FLT_TEST_1_t* _m, __CoderDbcCanFrame_t__* cframe)
 {
-  uint8_t i; for (i = 0; (i < FLT_TEST_1_DLC) && (i < 8); cframe->Data[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(FLT_TEST_1_DLC); cframe->Data[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_AUTO_ROLL
   _m->RC = (_m->RC + 1) & (0x0FU);
 #endif // TESTDB_AUTO_ROLL
 
 #ifdef TESTDB_AUTO_CSM
-  _m->CS = 0U;
+  _m->CS = (uint8_t) 0;
 #endif // TESTDB_AUTO_CSM
 
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->INT_TEST_2_ro = TESTDB_INT_TEST_2_ro_toS(_m->INT_TEST_2_phys);
-  _m->Accel_ro = TESTDB_Accel_ro_toS(_m->Accel_phys);
-  _m->FLT4_TEST_1_ro = TESTDB_FLT4_TEST_1_ro_toS(_m->FLT4_TEST_1_phys);
-  _m->FLT4_TEST_2_ro = TESTDB_FLT4_TEST_2_ro_toS(_m->FLT4_TEST_2_phys);
-  _m->FLT4_TEST_3_ro = TESTDB_FLT4_TEST_3_ro_toS(_m->FLT4_TEST_3_phys);
-  _m->INT_TEST_1_ro = TESTDB_INT_TEST_1_ro_toS(_m->INT_TEST_1_phys);
+  _m->INT_TEST_2_ro = (int8_t) TESTDB_INT_TEST_2_ro_toS(_m->INT_TEST_2_phys);
+  _m->Accel_ro = (uint16_t) TESTDB_Accel_ro_toS(_m->Accel_phys);
+  _m->FLT4_TEST_1_ro = (uint8_t) TESTDB_FLT4_TEST_1_ro_toS(_m->FLT4_TEST_1_phys);
+  _m->FLT4_TEST_2_ro = (uint8_t) TESTDB_FLT4_TEST_2_ro_toS(_m->FLT4_TEST_2_phys);
+  _m->FLT4_TEST_3_ro = (uint8_t) TESTDB_FLT4_TEST_3_ro_toS(_m->FLT4_TEST_3_phys);
+  _m->INT_TEST_1_ro = (uint8_t) TESTDB_INT_TEST_1_ro_toS(_m->INT_TEST_1_phys);
 #endif // TESTDB_USE_SIGFLOAT
 
-  cframe->Data[0] |= (_m->ValTable & (0x03U)) | ((_m->Position & (0x0FU)) << 4);
-  cframe->Data[1] |= ((_m->INT_TEST_2_ro & (0x7FU)) << 1);
-  cframe->Data[2] |= (_m->RC & (0x0FU)) | ((_m->CS & (0x0FU)) << 4);
-  cframe->Data[3] |= (_m->Accel_ro & (0xFFU));
-  cframe->Data[4] |= ((_m->Accel_ro >> 8) & (0x0FU)) | ((_m->FLT4_TEST_1_ro & (0x0FU)) << 4);
-  cframe->Data[5] |= (_m->FLT4_TEST_2_ro & (0x0FU)) | ((_m->FLT4_TEST_3_ro & (0x0FU)) << 4);
-  cframe->Data[6] |= ((_m->INT_TEST_1_ro & (0x0FU)) << 2);
+  cframe->Data[0] |= (uint8_t) ( (_m->ValTable & (0x03U)) | ((_m->Position & (0x0FU)) << 4U) );
+  cframe->Data[1] |= (uint8_t) ( ((_m->INT_TEST_2_ro & (0x7FU)) << 1U) );
+  cframe->Data[2] |= (uint8_t) ( (_m->RC & (0x0FU)) | ((_m->CS & (0x0FU)) << 4U) );
+  cframe->Data[3] |= (uint8_t) ( (_m->Accel_ro & (0xFFU)) );
+  cframe->Data[4] |= (uint8_t) ( ((_m->Accel_ro >> 8U) & (0x0FU)) | ((_m->FLT4_TEST_1_ro & (0x0FU)) << 4U) );
+  cframe->Data[5] |= (uint8_t) ( (_m->FLT4_TEST_2_ro & (0x0FU)) | ((_m->FLT4_TEST_3_ro & (0x0FU)) << 4U) );
+  cframe->Data[6] |= (uint8_t) ( ((_m->INT_TEST_1_ro & (0x0FU)) << 2U) );
 
 #ifdef TESTDB_AUTO_CSM
   _m->CS = ((uint8_t)GetFrameHash(cframe->Data, FLT_TEST_1_DLC, FLT_TEST_1_CANID, kXor4, 1));
-  cframe->Data[2] |= ((_m->CS & (0x0FU)) << 4);
+  cframe->Data[2] |= (uint8_t) ( ((_m->CS & (0x0FU)) << 4U) );
 #endif // TESTDB_AUTO_CSM
 
-  cframe->MsgId = FLT_TEST_1_CANID;
-  cframe->DLC = FLT_TEST_1_DLC;
-  cframe->IDE = FLT_TEST_1_IDE;
+  cframe->MsgId = (uint32_t) FLT_TEST_1_CANID;
+  cframe->DLC = (uint8_t) FLT_TEST_1_DLC;
+  cframe->IDE = (uint8_t) FLT_TEST_1_IDE;
   return FLT_TEST_1_CANID;
 }
 
@@ -355,40 +355,40 @@ uint32_t Pack_FLT_TEST_1_testdb(FLT_TEST_1_t* _m, __CoderDbcCanFrame_t__* cframe
 
 uint32_t Pack_FLT_TEST_1_testdb(FLT_TEST_1_t* _m, uint8_t* _d, uint8_t* _len, uint8_t* _ide)
 {
-  uint8_t i; for (i = 0; (i < FLT_TEST_1_DLC) && (i < 8); _d[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(FLT_TEST_1_DLC); _d[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_AUTO_ROLL
   _m->RC = (_m->RC + 1) & (0x0FU);
 #endif // TESTDB_AUTO_ROLL
 
 #ifdef TESTDB_AUTO_CSM
-  _m->CS = 0U;
+  _m->CS = (uint8_t) 0;
 #endif // TESTDB_AUTO_CSM
 
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->INT_TEST_2_ro = TESTDB_INT_TEST_2_ro_toS(_m->INT_TEST_2_phys);
-  _m->Accel_ro = TESTDB_Accel_ro_toS(_m->Accel_phys);
-  _m->FLT4_TEST_1_ro = TESTDB_FLT4_TEST_1_ro_toS(_m->FLT4_TEST_1_phys);
-  _m->FLT4_TEST_2_ro = TESTDB_FLT4_TEST_2_ro_toS(_m->FLT4_TEST_2_phys);
-  _m->FLT4_TEST_3_ro = TESTDB_FLT4_TEST_3_ro_toS(_m->FLT4_TEST_3_phys);
-  _m->INT_TEST_1_ro = TESTDB_INT_TEST_1_ro_toS(_m->INT_TEST_1_phys);
+  _m->INT_TEST_2_ro = (int8_t) TESTDB_INT_TEST_2_ro_toS(_m->INT_TEST_2_phys);
+  _m->Accel_ro = (uint16_t) TESTDB_Accel_ro_toS(_m->Accel_phys);
+  _m->FLT4_TEST_1_ro = (uint8_t) TESTDB_FLT4_TEST_1_ro_toS(_m->FLT4_TEST_1_phys);
+  _m->FLT4_TEST_2_ro = (uint8_t) TESTDB_FLT4_TEST_2_ro_toS(_m->FLT4_TEST_2_phys);
+  _m->FLT4_TEST_3_ro = (uint8_t) TESTDB_FLT4_TEST_3_ro_toS(_m->FLT4_TEST_3_phys);
+  _m->INT_TEST_1_ro = (uint8_t) TESTDB_INT_TEST_1_ro_toS(_m->INT_TEST_1_phys);
 #endif // TESTDB_USE_SIGFLOAT
 
-  _d[0] |= (_m->ValTable & (0x03U)) | ((_m->Position & (0x0FU)) << 4);
-  _d[1] |= ((_m->INT_TEST_2_ro & (0x7FU)) << 1);
-  _d[2] |= (_m->RC & (0x0FU)) | ((_m->CS & (0x0FU)) << 4);
-  _d[3] |= (_m->Accel_ro & (0xFFU));
-  _d[4] |= ((_m->Accel_ro >> 8) & (0x0FU)) | ((_m->FLT4_TEST_1_ro & (0x0FU)) << 4);
-  _d[5] |= (_m->FLT4_TEST_2_ro & (0x0FU)) | ((_m->FLT4_TEST_3_ro & (0x0FU)) << 4);
-  _d[6] |= ((_m->INT_TEST_1_ro & (0x0FU)) << 2);
+  _d[0] |= (uint8_t) ( (_m->ValTable & (0x03U)) | ((_m->Position & (0x0FU)) << 4U) );
+  _d[1] |= (uint8_t) ( ((_m->INT_TEST_2_ro & (0x7FU)) << 1U) );
+  _d[2] |= (uint8_t) ( (_m->RC & (0x0FU)) | ((_m->CS & (0x0FU)) << 4U) );
+  _d[3] |= (uint8_t) ( (_m->Accel_ro & (0xFFU)) );
+  _d[4] |= (uint8_t) ( ((_m->Accel_ro >> 8U) & (0x0FU)) | ((_m->FLT4_TEST_1_ro & (0x0FU)) << 4U) );
+  _d[5] |= (uint8_t) ( (_m->FLT4_TEST_2_ro & (0x0FU)) | ((_m->FLT4_TEST_3_ro & (0x0FU)) << 4U) );
+  _d[6] |= (uint8_t) ( ((_m->INT_TEST_1_ro & (0x0FU)) << 2U) );
 
 #ifdef TESTDB_AUTO_CSM
   _m->CS = ((uint8_t)GetFrameHash(_d, FLT_TEST_1_DLC, FLT_TEST_1_CANID, kXor4, 1));
-  _d[2] |= ((_m->CS & (0x0FU)) << 4);
+  _d[2] |= (uint8_t) ( ((_m->CS & (0x0FU)) << 4U) );
 #endif // TESTDB_AUTO_CSM
 
-  *_len = FLT_TEST_1_DLC;
-  *_ide = FLT_TEST_1_IDE;
+  *_len = (uint8_t) FLT_TEST_1_DLC;
+  *_ide = (uint8_t) FLT_TEST_1_IDE;
   return FLT_TEST_1_CANID;
 }
 
@@ -397,29 +397,29 @@ uint32_t Pack_FLT_TEST_1_testdb(FLT_TEST_1_t* _m, uint8_t* _d, uint8_t* _len, ui
 uint32_t Unpack_SIG_TEST_1_testdb(SIG_TEST_1_t* _m, const uint8_t* _d, uint8_t dlc_)
 {
   (void)dlc_;
-  _m->sig15_ro = __ext_sig__(( ((_d[1] & (0x7FU)) << 8) | (_d[0] & (0xFFU)) ), 15);
+  _m->sig15_ro = (int16_t) __ext_sig__(( ((_d[1] & (0x7FU)) << 8U) | (_d[0] & (0xFFU)) ), 15);
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->sig15_phys = TESTDB_sig15_ro_fromS(_m->sig15_ro);
+  _m->sig15_phys = (int32_t) TESTDB_sig15_ro_fromS(_m->sig15_ro);
 #endif // TESTDB_USE_SIGFLOAT
 
-  _m->sig15_2_ro = __ext_sig__(( ((_d[3] & (0x7FU)) << 8) | (_d[2] & (0xFFU)) ), 15);
+  _m->sig15_2_ro = (int16_t) __ext_sig__(( ((_d[3] & (0x7FU)) << 8U) | (_d[2] & (0xFFU)) ), 15);
 #ifdef TESTDB_USE_SIGFLOAT
   _m->sig15_2_phys = (sigfloat_t)(TESTDB_sig15_2_ro_fromS(_m->sig15_2_ro));
 #endif // TESTDB_USE_SIGFLOAT
 
-  _m->sig8_ro = __ext_sig__(( (_d[4] & (0xFFU)) ), 8);
+  _m->sig8_ro = (int8_t) __ext_sig__(( (_d[4] & (0xFFU)) ), 8);
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->sig8_phys = TESTDB_sig8_ro_fromS(_m->sig8_ro);
+  _m->sig8_phys = (int16_t) TESTDB_sig8_ro_fromS(_m->sig8_ro);
 #endif // TESTDB_USE_SIGFLOAT
 
-  _m->sig_7_ro = __ext_sig__(( (_d[5] & (0x7FU)) ), 7);
+  _m->sig_7_ro = (int8_t) __ext_sig__(( (_d[5] & (0x7FU)) ), 7);
 #ifdef TESTDB_USE_SIGFLOAT
   _m->sig_7_phys = (sigfloat_t)(TESTDB_sig_7_ro_fromS(_m->sig_7_ro));
 #endif // TESTDB_USE_SIGFLOAT
 
-  _m->U7_TEST_1_ro = ((_d[6] >> 1) & (0x7FU));
+  _m->U7_TEST_1_ro = (uint8_t) ( ((_d[6] >> 1U) & (0x7FU)) );
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->U7_TEST_1_phys = TESTDB_U7_TEST_1_ro_fromS(_m->U7_TEST_1_ro);
+  _m->U7_TEST_1_phys = (int16_t) TESTDB_U7_TEST_1_ro_fromS(_m->U7_TEST_1_ro);
 #endif // TESTDB_USE_SIGFLOAT
 
 #ifdef TESTDB_USE_DIAG_MONITORS
@@ -437,27 +437,27 @@ uint32_t Unpack_SIG_TEST_1_testdb(SIG_TEST_1_t* _m, const uint8_t* _d, uint8_t d
 
 uint32_t Pack_SIG_TEST_1_testdb(SIG_TEST_1_t* _m, __CoderDbcCanFrame_t__* cframe)
 {
-  uint8_t i; for (i = 0; (i < SIG_TEST_1_DLC) && (i < 8); cframe->Data[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(SIG_TEST_1_DLC); cframe->Data[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->sig15_ro = TESTDB_sig15_ro_toS(_m->sig15_phys);
-  _m->sig15_2_ro = TESTDB_sig15_2_ro_toS(_m->sig15_2_phys);
-  _m->sig8_ro = TESTDB_sig8_ro_toS(_m->sig8_phys);
-  _m->sig_7_ro = TESTDB_sig_7_ro_toS(_m->sig_7_phys);
-  _m->U7_TEST_1_ro = TESTDB_U7_TEST_1_ro_toS(_m->U7_TEST_1_phys);
+  _m->sig15_ro = (int16_t) TESTDB_sig15_ro_toS(_m->sig15_phys);
+  _m->sig15_2_ro = (int16_t) TESTDB_sig15_2_ro_toS(_m->sig15_2_phys);
+  _m->sig8_ro = (int8_t) TESTDB_sig8_ro_toS(_m->sig8_phys);
+  _m->sig_7_ro = (int8_t) TESTDB_sig_7_ro_toS(_m->sig_7_phys);
+  _m->U7_TEST_1_ro = (uint8_t) TESTDB_U7_TEST_1_ro_toS(_m->U7_TEST_1_phys);
 #endif // TESTDB_USE_SIGFLOAT
 
-  cframe->Data[0] |= (_m->sig15_ro & (0xFFU));
-  cframe->Data[1] |= ((_m->sig15_ro >> 8) & (0x7FU));
-  cframe->Data[2] |= (_m->sig15_2_ro & (0xFFU));
-  cframe->Data[3] |= ((_m->sig15_2_ro >> 8) & (0x7FU));
-  cframe->Data[4] |= (_m->sig8_ro & (0xFFU));
-  cframe->Data[5] |= (_m->sig_7_ro & (0x7FU));
-  cframe->Data[6] |= ((_m->U7_TEST_1_ro & (0x7FU)) << 1);
+  cframe->Data[0] |= (uint8_t) ( (_m->sig15_ro & (0xFFU)) );
+  cframe->Data[1] |= (uint8_t) ( ((_m->sig15_ro >> 8U) & (0x7FU)) );
+  cframe->Data[2] |= (uint8_t) ( (_m->sig15_2_ro & (0xFFU)) );
+  cframe->Data[3] |= (uint8_t) ( ((_m->sig15_2_ro >> 8U) & (0x7FU)) );
+  cframe->Data[4] |= (uint8_t) ( (_m->sig8_ro & (0xFFU)) );
+  cframe->Data[5] |= (uint8_t) ( (_m->sig_7_ro & (0x7FU)) );
+  cframe->Data[6] |= (uint8_t) ( ((_m->U7_TEST_1_ro & (0x7FU)) << 1U) );
 
-  cframe->MsgId = SIG_TEST_1_CANID;
-  cframe->DLC = SIG_TEST_1_DLC;
-  cframe->IDE = SIG_TEST_1_IDE;
+  cframe->MsgId = (uint32_t) SIG_TEST_1_CANID;
+  cframe->DLC = (uint8_t) SIG_TEST_1_DLC;
+  cframe->IDE = (uint8_t) SIG_TEST_1_IDE;
   return SIG_TEST_1_CANID;
 }
 
@@ -465,26 +465,26 @@ uint32_t Pack_SIG_TEST_1_testdb(SIG_TEST_1_t* _m, __CoderDbcCanFrame_t__* cframe
 
 uint32_t Pack_SIG_TEST_1_testdb(SIG_TEST_1_t* _m, uint8_t* _d, uint8_t* _len, uint8_t* _ide)
 {
-  uint8_t i; for (i = 0; (i < SIG_TEST_1_DLC) && (i < 8); _d[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(SIG_TEST_1_DLC); _d[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_USE_SIGFLOAT
-  _m->sig15_ro = TESTDB_sig15_ro_toS(_m->sig15_phys);
-  _m->sig15_2_ro = TESTDB_sig15_2_ro_toS(_m->sig15_2_phys);
-  _m->sig8_ro = TESTDB_sig8_ro_toS(_m->sig8_phys);
-  _m->sig_7_ro = TESTDB_sig_7_ro_toS(_m->sig_7_phys);
-  _m->U7_TEST_1_ro = TESTDB_U7_TEST_1_ro_toS(_m->U7_TEST_1_phys);
+  _m->sig15_ro = (int16_t) TESTDB_sig15_ro_toS(_m->sig15_phys);
+  _m->sig15_2_ro = (int16_t) TESTDB_sig15_2_ro_toS(_m->sig15_2_phys);
+  _m->sig8_ro = (int8_t) TESTDB_sig8_ro_toS(_m->sig8_phys);
+  _m->sig_7_ro = (int8_t) TESTDB_sig_7_ro_toS(_m->sig_7_phys);
+  _m->U7_TEST_1_ro = (uint8_t) TESTDB_U7_TEST_1_ro_toS(_m->U7_TEST_1_phys);
 #endif // TESTDB_USE_SIGFLOAT
 
-  _d[0] |= (_m->sig15_ro & (0xFFU));
-  _d[1] |= ((_m->sig15_ro >> 8) & (0x7FU));
-  _d[2] |= (_m->sig15_2_ro & (0xFFU));
-  _d[3] |= ((_m->sig15_2_ro >> 8) & (0x7FU));
-  _d[4] |= (_m->sig8_ro & (0xFFU));
-  _d[5] |= (_m->sig_7_ro & (0x7FU));
-  _d[6] |= ((_m->U7_TEST_1_ro & (0x7FU)) << 1);
+  _d[0] |= (uint8_t) ( (_m->sig15_ro & (0xFFU)) );
+  _d[1] |= (uint8_t) ( ((_m->sig15_ro >> 8U) & (0x7FU)) );
+  _d[2] |= (uint8_t) ( (_m->sig15_2_ro & (0xFFU)) );
+  _d[3] |= (uint8_t) ( ((_m->sig15_2_ro >> 8U) & (0x7FU)) );
+  _d[4] |= (uint8_t) ( (_m->sig8_ro & (0xFFU)) );
+  _d[5] |= (uint8_t) ( (_m->sig_7_ro & (0x7FU)) );
+  _d[6] |= (uint8_t) ( ((_m->U7_TEST_1_ro & (0x7FU)) << 1U) );
 
-  *_len = SIG_TEST_1_DLC;
-  *_ide = SIG_TEST_1_IDE;
+  *_len = (uint8_t) SIG_TEST_1_DLC;
+  *_ide = (uint8_t) SIG_TEST_1_IDE;
   return SIG_TEST_1_CANID;
 }
 
@@ -493,8 +493,8 @@ uint32_t Pack_SIG_TEST_1_testdb(SIG_TEST_1_t* _m, uint8_t* _d, uint8_t* _len, ui
 uint32_t Unpack_EMPTY_EXT_ID_testdb(EMPTY_EXT_ID_t* _m, const uint8_t* _d, uint8_t dlc_)
 {
   (void)dlc_;
-  _m->ValTest = ((_d[0] >> 6) & (0x03U));
-  _m->CS = (_d[1] & (0x3FU));
+  _m->ValTest = (uint8_t) ( ((_d[0] >> 6U) & (0x03U)) );
+  _m->CS = (uint8_t) ( (_d[1] & (0x3FU)) );
 
 #ifdef TESTDB_USE_DIAG_MONITORS
   _m->mon1.dlc_error = (dlc_ < EMPTY_EXT_ID_DLC);
@@ -515,23 +515,23 @@ uint32_t Unpack_EMPTY_EXT_ID_testdb(EMPTY_EXT_ID_t* _m, const uint8_t* _d, uint8
 
 uint32_t Pack_EMPTY_EXT_ID_testdb(EMPTY_EXT_ID_t* _m, __CoderDbcCanFrame_t__* cframe)
 {
-  uint8_t i; for (i = 0; (i < EMPTY_EXT_ID_DLC) && (i < 8); cframe->Data[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(EMPTY_EXT_ID_DLC); cframe->Data[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_AUTO_CSM
-  _m->CS = 0U;
+  _m->CS = (uint8_t) 0;
 #endif // TESTDB_AUTO_CSM
 
-  cframe->Data[0] |= ((_m->ValTest & (0x03U)) << 6);
-  cframe->Data[1] |= (_m->CS & (0x3FU));
+  cframe->Data[0] |= (uint8_t) ( ((_m->ValTest & (0x03U)) << 6U) );
+  cframe->Data[1] |= (uint8_t) ( (_m->CS & (0x3FU)) );
 
 #ifdef TESTDB_AUTO_CSM
   _m->CS = ((uint8_t)GetFrameHash(cframe->Data, EMPTY_EXT_ID_DLC, EMPTY_EXT_ID_CANID, kXor8, 1));
-  cframe->Data[1] |= (_m->CS & (0x3FU));
+  cframe->Data[1] |= (uint8_t) ( (_m->CS & (0x3FU)) );
 #endif // TESTDB_AUTO_CSM
 
-  cframe->MsgId = EMPTY_EXT_ID_CANID;
-  cframe->DLC = EMPTY_EXT_ID_DLC;
-  cframe->IDE = EMPTY_EXT_ID_IDE;
+  cframe->MsgId = (uint32_t) EMPTY_EXT_ID_CANID;
+  cframe->DLC = (uint8_t) EMPTY_EXT_ID_DLC;
+  cframe->IDE = (uint8_t) EMPTY_EXT_ID_IDE;
   return EMPTY_EXT_ID_CANID;
 }
 
@@ -539,22 +539,22 @@ uint32_t Pack_EMPTY_EXT_ID_testdb(EMPTY_EXT_ID_t* _m, __CoderDbcCanFrame_t__* cf
 
 uint32_t Pack_EMPTY_EXT_ID_testdb(EMPTY_EXT_ID_t* _m, uint8_t* _d, uint8_t* _len, uint8_t* _ide)
 {
-  uint8_t i; for (i = 0; (i < EMPTY_EXT_ID_DLC) && (i < 8); _d[i++] = 0);
+  uint8_t i; for (i = 0u; i < TESTDB_VALIDATE_DLC(EMPTY_EXT_ID_DLC); _d[i++] = TESTDB_INITIAL_BYTE_VALUE);
 
 #ifdef TESTDB_AUTO_CSM
-  _m->CS = 0U;
+  _m->CS = (uint8_t) 0;
 #endif // TESTDB_AUTO_CSM
 
-  _d[0] |= ((_m->ValTest & (0x03U)) << 6);
-  _d[1] |= (_m->CS & (0x3FU));
+  _d[0] |= (uint8_t) ( ((_m->ValTest & (0x03U)) << 6U) );
+  _d[1] |= (uint8_t) ( (_m->CS & (0x3FU)) );
 
 #ifdef TESTDB_AUTO_CSM
   _m->CS = ((uint8_t)GetFrameHash(_d, EMPTY_EXT_ID_DLC, EMPTY_EXT_ID_CANID, kXor8, 1));
-  _d[1] |= (_m->CS & (0x3FU));
+  _d[1] |= (uint8_t) ( (_m->CS & (0x3FU)) );
 #endif // TESTDB_AUTO_CSM
 
-  *_len = EMPTY_EXT_ID_DLC;
-  *_ide = EMPTY_EXT_ID_IDE;
+  *_len = (uint8_t) EMPTY_EXT_ID_DLC;
+  *_ide = (uint8_t) EMPTY_EXT_ID_IDE;
   return EMPTY_EXT_ID_CANID;
 }